Fixing warnings

+++ ENgine/Editor/Editor.cpp
@@ -207,7 +207,7 @@ namespace Oak
 
 		ImVec2 viewportPos = ImVec2(io.MousePos.x - ImGui::GetCursorScreenPos().x, io.MousePos.y - ImGui::GetCursorScreenPos().y);
 
-		Oak::root.controls.OverrideMousePos(viewportPos.x, viewportPos.y);
+		Oak::root.controls.OverrideMousePos((int)viewportPos.x, (int)viewportPos.y);
 
 		ImVec2 size = ImGui::GetContentRegionAvail();
 		Oak::root.render.GetDevice()->SetBackBuffer(0, (int)size.x, (int)size.y, &hwnd);
